refactor: move 'diff' command to vue folder
