+++ packages/cli/src/commands/frameworks/index.ts
@@ -1,11 +1,13 @@
 import loadConfigVue from './vue/config'
 import initVue from './vue/init'
 import addVue from './vue/add'
+import diffVue from './vue/diff'
 
 export default {
   vue: {
     loadConfig: loadConfigVue,
     init: initVue,
     add: addVue,
+    diff: diffVue,
   },
 }
